rag_system_rebuild/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
langchain==0.3
langchain-community==0.3
langchain-openai==0.2
langchain-ollama==0.2
sentence-transformers==2.2.2
huggingface-hub>=0.19.0,<0.20.0
chromadb==0.4.18
python-multipart>=0.0.12,<0.1
pydantic>=2.1.0
jinja2>=3.1.2
python-dotenv==1.0.0
tiktoken>=0.1
requests==2.31.0
pytest==7.4.3
pytest-asyncio==0.21.1
pypdf>=3.0.0
PyPDF2>=3.0.0
langfuse>=2.0.0
